implementing Transaction class
I did some basic implementation in the Transaction class.

It's not working properly but I sent it anyways just to share what I've done until now.

Lastly, is it rational for the Transaction class to contain a list of Products? Every Transaction entry should have one Product at a time because each one has its own parameteres, place, enviroment, and a date. Even though the date might be the same for each entry, every entry should be different for the rest of the parameters.

So Transaction should take a single Product at a time?